admin post apis

+++ src/controllers/post.controller.ts
@@ -250,7 +250,16 @@ export const updatePost = TryCatch(
 export const deletePost = TryCatch(
   async (req: Request, res: Response, next: NextFunction) => {
     const id = req.params.id;
+    const user = req.user;
     const postId = parseInt(id);
+
+    const post = await prisma.post.findFirst({
+      where: { id: postId, authorId: user?.id },
+    });
+
+    if (!post) {
+      return next(new ErrorHandler(400, "You are not an author of this post."));
+    }
 
     await prisma.reply.deleteMany({
       where: {
@@ -264,6 +273,10 @@ export const deletePost = TryCatch(
       where: {
         postId,
       },
+    });
+
+    await prisma.like.deleteMany({
+      where: { postId },
     });
 
     await prisma.postCategory.deleteMany({
@@ -499,4 +512,117 @@ export const likedPost = TryCatch(
   }
 );
 
-//Note admin delete comment and reply
+//Admin
+export const getAllPostAdmin = TryCatch(
+  async (req: Request, res: Response, next: NextFunction) => {
+    const posts = await prisma.post.findMany({
+      include: {
+        categories: {
+          include: {
+            category: {
+              select: {
+                name: true,
+              },
+            },
+          },
+        },
+      },
+    });
+
+    res.status(200).json({
+      success: true,
+      posts,
+    });
+  }
+);
+
+export const deletePosts = TryCatch(
+  async (req: Request, res: Response, next: NextFunction) => {
+    const { postId } = req.body;
+
+    await prisma.reply.deleteMany({
+      where: {
+        comment: {
+          postId,
+        },
+      },
+    });
+
+    await prisma.comment.deleteMany({
+      where: {
+        postId,
+      },
+    });
+
+    await prisma.like.deleteMany({
+      where: { postId },
+    });
+
+    await prisma.postCategory.deleteMany({
+      where: { postId },
+    });
+
+    await prisma.post.delete({
+      where: {
+        id: postId,
+      },
+    });
+
+    res.status(200).json({
+      success: true,
+      message: "Post deleted successfully",
+    });
+  }
+);
+
+export const deleteComment = TryCatch(
+  async (req: Request, res: Response, next: NextFunction) => {
+    const { id } = req.body;
+
+    const comment = await prisma.comment.findUnique({
+      where: { id },
+    });
+
+    if (!comment) {
+      return next(new ErrorHandler(404, "Comment not found"));
+    }
+
+    await prisma.reply.deleteMany({
+      where: {
+        commentId: id,
+      },
+    });
+
+    await prisma.comment.delete({
+      where: { id },
+    });
+
+    res.status(200).json({
+      success: true,
+      message: "Comment deleted successfully",
+    });
+  }
+);
+
+export const deleteReply = TryCatch(
+  async (req: Request, res: Response, next: NextFunction) => {
+    const { id } = req.body;
+
+    const reply = await prisma.reply.findFirst({
+      where: { id },
+    });
+
+    if (!reply) {
+      return next(new ErrorHandler(404, "Reply not found"));
+    }
+
+    await prisma.reply.delete({
+      where: { id },
+    });
+
+    res.status(200).json({
+      success: true,
+      message: "Reply deleted successfully",
+    });
+  }
+);
